// src/lib/enums/product.enum.ts
export enum ProductSize {
    SINGLE = "SINGLE",
    DOUBLE = "DOUBLE",
    CORNER = "CORNER",
    SET = "SET"
}


export enum ProductStatus {
    PAUSE = "PAUSE",
    PROCESS = "PROCESS",
    SOLDOUT = "SOLDOUT",
    DELETE = "DELETE"
}

export enum ProductCollection {
    MODERN = "MODERN",
    CLASSIC = "CLASSIC",
    RETRO = "RETRO",
    FAMILY = "FAMILY",
    LUXURY = "LUXURY",
    OTHER = "OTHER"
}

export enum ProductMaterial {
    LEATHER = "LEATHER",
    FABRIC = "FABRIC",
    VELVET = "VELVET",
    WOOD = "WOOD"
}